v4/requirements.txt
ccxt>=4.0.0
rich>=13.0.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
pyarrow>=12.0.0  # For parquet caching
